add new component SignUpFrom, separate logic and types in form
