+++ src/components/ui/sign-up-form/SignUpForm.interface.ts
@@ -1,0 +1,22 @@
+export interface UserData {
+    email: string;
+    password: string;
+    fullName: string;
+    firstName: string;
+    lastName: string;
+    phone: string;
+    picture: string;
+    birthDate: string;
+}
+
+export interface CreateUserRequest {
+    email: string;
+    password: string;
+    is_active: boolean | null;
+    is_superuser: boolean | null;
+    is_verified: boolean | null;
+    full_name: string;
+    phone: string;
+    picture: string;
+    birth_date: string;
+}
