+++ src/types/types.ts
@@ -27,25 +27,3 @@ export interface OptionRate {
 }
 
 
-export interface UserData {
-    email: string;
-    password: string;
-    fullName: string;
-    firstName: string;
-    lastName: string;
-    phone: string;
-    picture: string;
-    birthDate: string;
-}
-
-export interface CreateUserRequest {
-    email: string;
-    password: string;
-    is_active: boolean | null;
-    is_superuser: boolean | null;
-    is_verified: boolean | null;
-    full_name: string;
-    phone: string;
-    picture: string;
-    birth_date: string;
-}
